Reverted to working state
I forgot to change AXI data width from 1024 to 512 in the python script.
For that reason every 2nd FIFO item didn't show. I didn't notice that
and I thought that incorrect results are due to timing issues caused by
combinational logic and spent whole day trying to use more clocked
logic. This commit reverts the state to the following commit:
4ab48d53302afa3bd600c37e92bc6c6612f1f17d

But it keeps all the commits created in-between just in case if timing
issues appeared in future and the code would have to be modified in
similar fashion.